I've modified all of the getter methods to use the approriate ByteBuffer.get*() call.  For 1-byte buffers, call get(), for 2, call getShort(), for 4 call getInt(), etc. and cast all of these to an int for return.
Also, I've modified getChecksum() so it correctly returns the two checsum bytes as an int (bit jockeying was required).

I modified addToByteSum so it now handles byte by byte (vs += shortValue if a short is passed in).

Lastly, I removed the call to addToByteSum() for the checksum field itself - it's not part of the stated checksum.

Oh - created the ensembleByteCount field for tracking bytes.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@177 9e959af2-1677-4b95-8350-622619c7fe3b