I've modified the execute() method with changes that deal with situations where a header id sequence (0x7F7F) is immediately preceded by 0x7F to create a sequence of 0x7F7F7F.  The parsing logic was treating the first two 7F bytes as the next header id incorrectly, and was causing the state machine to permanently stay in state 4.
The new code reads *exactly* ensembleBytes + 4 forward in the stream (two checksum bytes + the two header bytes of the next ensemble) and verifies the checksum before rewinding to the start of the next ensemble header and starting the parsing over again.  Note: After reading the second 0x7F of the next header id, we don't add it to the ensembleByteCount, and so evaluation statement tests for 'ensembleBytes + 3', not 4.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@220 9e959af2-1677-4b95-8350-622619c7fe3b