I've added fixes to the execute() method in order for it to successfully write commands ( and receive the results) from the Seabird instrument.  There were two main issues:
1) I was sending Unix <LF> line endings at the end of commands rather than Windows <CR><LF> line endings

2) The code was sending commands too quickly, and I've added a short wait time between commands being sent

I've also realized that the first command response often comes back only partially, because the instrument needs to be awakened, and so I send an initial ID? command just to get the instrument fully responding.

I've also changed the 'socket' variable to 'socketChannel', which is more accurate, and have updated the logging output to make the rbnb flushes more apparent.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@266 9e959af2-1677-4b95-8350-622619c7fe3b