I've modified the FileArchiverSink in order to support two things:
1) multiple data types to be archived (not just bytearray)
2) 3 archive interval (hourly, daily, weekly)

In doing so, I've modified the main() method to get the archive interval dynamically.  I've modified the setArchiveTime() method to handle the three scenarios.  I've added a new command line option (-I) to state which archive interval is desired.

For the data type support, I've used ChannelMap.GetData() in the exportData() method so it will cast any incoming data from the DataTurbine to a byte[], and so we can handle all of the data types equally.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@281 9e959af2-1677-4b95-8350-622619c7fe3b