I've made some significant changes to the TChainSource driver in order to make it more flexible in terms of handling different sample kine endings, delimiter characters, and logging, etc.  New fields include:
1) DEFAULT_LOG_CONFIGURATION_FILE: The default log configuration file location

2) logConfigurationFile: The log configuration file location

3) DEFAULT_FIELD_DELIMITER: The default field delimiter string used to separate data fields in the sample

4) fieldDelimiter: The field delimiter string used to separate data fields in the sample

5) DEFAULT_FIRST_DELIMITER_BYTE: The default first byte of the sample delimiter

6) DEFAULT_SECOND_DELIMITER_BYTE: The default second byte of the sample delimiter

7) firstDelimiterByte: The first byte of the sample delimiter

8) secondDelimiterByte: The default second byte of the sample delimiter

With these new fields available, we can now set the field delimiter and record delimiters via the command line arguments.  The setOptions() method and the setArgs() methods were modified accordingly, with the added -d and -l command line flags.

The execute() method was modified to use these values to identify the sample record delimiter, and to use the appropriate number of field delimiter characters (usually a number of spaces) before adding in the timestamp to the sample record.

Two methds were added:

getFieldDelimiter(): A method that gets the field delimiter string, which is used to append a timestamp at the end of the sample using the correct delimiter.

getLogConfigurationFile(): A method that gets the log configuration file location

I've modified the Log4J logging code to use a PropertyConfigurator that reads settings from the log4j.properties file, and sets the appropriate log level.




git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@295 9e959af2-1677-4b95-8350-622619c7fe3b