I'm adding a new Source driver for the WetLabs ECO FLNTU fluorometry and turbidity sensor.  The instrument identifier will be FLNT in the naming conventions.  This driver is similar to the TChainSource driver, except that it doesn't need to append a timestamp, and it must recognize 'sets' of samples.
The FLNT samples at predetermined intervals, with predetermined pauses between intervals.  Beginnings and endings of sample sets are delimited by 'mvs 1' and 'mvs 0'.  Sample lines of data between these markers are flushed to the data turbine one at a time, keeping the data coming in in as real-time as possible.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@304 9e959af2-1677-4b95-8350-622619c7fe3b