This is a significant number of changes to the Davis weather station parser.  I've added in a new field called packetBuffer, which stores the binary LOOP packet passed into the parser for all of the getter methods to access if need be.
Next, I've modified the constructor to make use of this internal packetBuffer field.

Then, I've modified each of the getter methods for each field, changing the returned data type to the appropriate data type for each field (strings are output as strings, integers as integers, rather than the underlying ByteBuffer).

Lastly, I've updated the main() method with calls to each of the getter methods for testing, and changed the method such that it takes the path to a binary LOOP file as an argument.  The LOOP file should only contain a single LOOP packet for parse testing purposes.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@368 9e959af2-1677-4b95-8350-622619c7fe3b