This commit is for major changes to the execute() method:
1) I've added in two ChannelMap objects, one for registering ADAM channels with the DAtaTurbine, and the second for flushing the voltage values to the DT for each of the registered channels.

2) For some reason the call to DatagramPacket.getAddress() returns an IP address preceded by a slash (/).  I've trimmed off the slash to get just the IP.

3) To determine which ADAM voltage channels represent which measurements in each Kilo Nalu node, I've built an XML configuration file that stores the mappings (lib/sensor.properties.xml).  In execute(), we now use the Apache XMLConfiguration object to read the contents of the config file.  We then create a HashMap to store a hash between the voltage channel (as an Integer) and the RBNB channel path associated with that voltage channel.  The RBNB channel path is built from information found in the XML config file.

As each datagram is read off of the wire, we iterate through the <sensor> definitions in the XML config file, find the definition that matches the IP address of the datagram, and then build the RBNB channel paths, and put them into the HashMap with their corresponding voltage channel.

Once the HashMap is complete (for the given IP address), we populate the RBNB channelMap with the channel path and the voltage value, using a call to the AdamParser (like adamParser.getChannelZero()).

The channels are registered with the DataTurbine, and then Flushed to the DataTurbine with the current timestamp.

4) I've added in support for the log4j PropertyConfigurator like the rest of the instrument source classes.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@436 9e959af2-1677-4b95-8350-622619c7fe3b