Due to the nature of the ADAM module engineering data streams, which are sent as UDP datagrams to a single port on a single server, only one client can  bind to the port at a time to receive the datagrams.  Likewise, since each RBNB Source driver that extends RBNBSource and RBNBBase can represent one and only one Source object, a mechanism was needed to proxy incoming datagrams and send those datagrams to multiple AdamSource drivers based on the incoming IP address stated in the datagram.  AdamDispatcher provides this proxy service.  It provides the standard fields and methods needed to gain configuration information from the command line:
DEFAULT_SOURCE_HOST_NAME &
sourceHostName -  The domain name or IP address of the host machine
                                    that this Source represents and from which the data 
                                    will stream. 

DEFAULT_SOURCE_HOST_PORT &
sourceHostPort - The UDP port to connect to on the Source host machine

DEFAULT_SERVER_NAME &
serverName - The IP address or DNS name of the RBNB server

DEFAULT_SERVER_PORT &
serverPort - The TCP port of the RBNB server

DEFAULT_ARCHIVE_MODE &
archiveMode - The archive mode for RBNB Source clients


Fields to store XML configuration information (using the Apache Commons Configuration library):

xmlConfigurationFile - The XML configuration file location for the list of sensor properties

Fields for Log4J logging:

DEFAULT_LOG_CONFIGURATION_FILE &
logConfigurationFile - The log configuration file location

Fields for UDP communication:

datagramSocket -  The socket used to establish UDP communication with the instrument
datagramPacket - The datagram object used to represent an individual incoming UDP packet
connected - The the connection status for the UDP port for the data streams
streamingThread - The thread that is run for streaming data from the instrument
readyToStream - A boolean field indicating if the instrument connection is ready to stream
RETRY_INTERVAL - An internal Thread setting used to specify how long, in milliseconds, 
                                       the execution of the data streaming Thread should wait before re-executing

And a hash map used to hold Source to IP mappings:

sourceMap - A hash map that contains IP address to RBNB Source mappings

The following methods provide instantiation and both UDP (from the instrument) and TCP (to the DataTurbine) communication:

AdamDispatcher() - Constructor - create an empty instance of the AdamDispatcher object
AdamDispatcher() - Constructor - creates an instance of the AdamDispatcher object using
                                    configuration information from the command line and XML configuration file

 connect()  - A method used to connect to the UDP port of the host that will have the UDP stream of data packets,
                      and that will also connect each of the AdamSource drivers to the DataTurbine

disconnect()  - A method used to disconnect from the UDP port of the host that has the UDP stream 
                          of data packets, and that will also disconnect each of the AdamSource drivers from the DataTurbine.

execute()  - A method that executes the streaming of data from the source to the RBNB server after all 
                     configuration of settings, connections to hosts, and thread initiatizing occurs.

getBufferSize()  - A method that sets the size, in bytes, of the ByteBuffer used in streaming data from a 
                               source instrument via a TCP connection

getHostName()  - A method that returns the domain name or IP address of the source instrument

getHostPort()  - A method that returns the TCP port of the source instrument

getLogConfigurationFile()  - A method that gets the log configuration file location

isConnected()  - A method that returns the connection status of the UDP socket and RBNB Sources

isRunning()  -  A method that returns true if the UDP DatagramSocket connection is established
                           and if the data streaming Thread has been started

main() - The main method for running the code

printUsage() - Print out the usage of this application to standard output.

setArgs() - A method that sets the command line arguments for this class.

setHostName() - A method that sets the domain name or IP address of the UDP stream host

setHostPort() - A method that sets the UDP port of the UDP stream host

setLogConfigurationFile() - A method that sets the log configuration file name

setOptions() - A method that sets the command line options for this class.

setServerName() - A method that sets the domain name or IP address of the RBNB server

setServerPort() - A method that sets the TCP port of the RBNB server

start() - A method that starts the streaming of data from the UDP host to the RBNB server

stop()  - A method that stops the streaming of data between the UDP host and the RBNB server

The class generally gets configuration information from the command line and the sensors.properties.xml file, establishes communication over a UDP socket to the server (port 5168), and establishes TCP communication of multiple AdamSource objects with the DataTurbine (one per sensor in the XML config file).  Once connections are made, the execute() method recieves UDP packets, parses out the IP address from the packet, determines which RBNB source it belongs to (via a hash map lookup), and then calls AdamSource.process() to send the data packet to the DataTurbine.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@443 9e959af2-1677-4b95-8350-622619c7fe3b