I'm adding a CTDConverter class, which effectively converts raw decimal CTD frequencies and voltages to meaningful engineering units such as temperature in degrees celsius, conductivity in Siemens/m, and pressure in decibars.
The class will use log4J for logging, and contains the DEFAULT_LOG_CONFIGURATION_FILE field and the logConfigurationFile field to get logging configurations.

It also uses the Apache Commons RealMatrix interface and the API to store a matrix of double values (both converted and raw).  See the convertedDataValuesMatrix and dataValuesMatrix fields.

The constructor takes an instance of the CTDParser class which stores all of the metadata and data fields that were parsed from the instrument output file.  It specifically contains the matrix of raw frquencies and voltages that will be converted, and the various temperature, conductivity, and pressure coefficients needed to do the conversions.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@471 9e959af2-1677-4b95-8350-622619c7fe3b