I've added in the SeahorseSource class used to harvest a compressed hexadecimal ASCII data file from a Brooke Ocean Seahorse profiler equipped with a Seacat SBE19plus CTD. Communication is over a TCP socket connection with an Iridium satellite modem and a serial2ip converter host. The data samples are then converted into RBNB frames and pushed into the RBNB DataTurbine real time server. This class extends org.nees.rbnb.RBNBSource, which in turn extends org.nees.rbnb.RBNBBase, and therefore follows the API conventions found in the org.nees.rbnb code.
The parsing of the data stream relies on the premise that each sample (scan) of data is a Hex-encoded string of values, and that each sample is terminated by a newline character (\n). It is also assumed that the sample rate is 4Hz, or a sample is taken every 15 seconds.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@504 9e959af2-1677-4b95-8350-622619c7fe3b