I've added the following class variables that are used for communication between the SeahorseSource class and the DataTurbine, and the SeahorseSource and the instrument:
DEFAULT_ARCHIVE_MODE           - A default archive mode for the given source connection to the RBNB server.  
archiveMode                    - The mode in which the source interacts with the RBNB archive. Valid modes include 'append', 'create', 'load' and 'none', however, Kilo Nalu instruments should append to an archive, which will create one if none exist.
DEFAULT_BUFFER_SIZE            - The default size of the ByteBuffer used to beffer the TCP stream from the source instrument.
DEFAULT_RBNB_CHANNEL           - A default RBNB channel name for the given source instrument
rbnbChannelName                - The name of the RBNB channel for this data stream  
DEFAULT_SOURCE_HOST_NAME       - A default source IP address for the given source instrument represents and from which the data will stream.                                                  - The domain name or IP address of the host machine that this Source 
DEFAULT_SOURCE_HOST_PORT       - A default source TCP port for the given source instrument
sourceHostPort                 - The TCP port to connect to on the Source host machine 
socketChannel                  - The socket channel used to establish TCP communication with the instrument  
resultByteCount                - The number of bytes in the ensemble as each byte is read from the stream  


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@505 9e959af2-1677-4b95-8350-622619c7fe3b