To handle XML-based metadata returned from the various Get* commands for the CTDs, I've added a number of fields to the CTDParser class and an overridden version of the setMetadata() method.  I've also added an empty constructor which should be used to build the metadata fields when they are derived from CTD commands with XML-based return output.
I've removed the xmlConfiguration instance field since we are using the NCEAS XML Utilities library to handle XML parsing rather than the Apache XMLConfiguration class.

The new fields include:

synchronizationMode -                          A field that stores the CTD synchronization mode as a String

SYNCHRONIZATION_MODE -            A field that stores the CTD synchronization mode key as a String

deviceType -                                            A field that stores the instrument device type as a String

DEVICE_TYPE -                                     A field that stores the instrument device type key as a String

instrumentSerialNumber -                    A field that stores the instrument serial number as a String */

INSTRUMENT_SERIAL_NUMBER - A field that stores the instrument serial number key as a String

pressureInstalled -                                 A field that stores the pressure installed value as a String

PRESSURE_INSTALLED -                 A field that stores the pressure installed key as a String

pumpInstalled -                                      A field that stores the pump installed value as a String

PUMP_INSTALLED -                           A field that stores the pump installed key as a String

hasPump -                                              A boolean field indicating if a pump is present on the instrument

outputSalinity -                                      A field that stores the output salinity state as a String

OUTPUT_SALINITY -                          A field that stores the output salinity state key as a String  

outputSoundVelocity -                         A field that stores the output sound velocity state as a String                      

OUTPUT_SOUND_VELOCITY -       A field that stores the output sound velocity state key as a String

xmlMetadata  -                                      An XML document object used to access CTD metadata reported in XML syntax

The setMetadata(xmlString) method takes an XML string as input and parses the metadata fields from the XML.  It currently handles the following root elements: ConfigurationData, StatusData, CallibrationCoefficients, EventCounters, HardwareData.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@547 9e959af2-1677-4b95-8350-622619c7fe3b