To handle the instrument XML CalibrationCoefficients output, I've added the following fields:
temperatureCalibrationFormat -                         A field that stores the temperature calibration format as a String

TEMPERATURE_CALIBRATION_FORMAT - A field that stores the temperature calibration format key as a String

conductivityCalibrationFormat -                         A field that stores the conductivity calibration format as a String

CONDUCTIVITY_CALIBRATION_FORMAT - A field that stores the conductivity calibration format key as a String

conductivityCoefficientWBOTC -                        A field that stores the conductivity coefficient WBOTC as a String

CONDUCTIVITY_COEFFICIENT_WBOTC -   A field that stores the conductivity coefficient WBOTC key as a String

pressureCalibrationDate -                                  A field that stores the pressure calibration date as a String

PRESSURE_CALIBRATION_DATE -              A field that stores the pressure calibration date key as a String

pressureCalibrationFormat -                              A field that stores the pressure calibration format as a String

PRESSURE_CALIBRATION_FORMAT -        A field that stores the pressure calibration format key as a String

And then I've modified the setMetadata(xmlString) method to parse the <CalibrationCoefficients> XML output and populate the appropriate fields.  This section iterates through the <Calibration> elements and tests if the current element has the 'Temprerature', 'Conductivity', or 'Pressure' id attributes and handles the metadata fields accordingly.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@549 9e959af2-1677-4b95-8350-622619c7fe3b