I'm modifying the CTDSource class to provide local serial port support for data streaming.  I've modified some of the documentation, but the largest changes include:
The addition of the following fields:

DEFAULT_SERIAL_PORT - The default serial port name used for serial communications

serialPortName - The serial port name used for serial communications 

isSerial - A boolean stating if the connection is serial vs socket

channel - The socket or file channel used for instrument communication


I've changed the execute() method to test for either a serial or socket connection type, and based on the type, call either getSocketConnection() or getSerialConnection().  I've also replaced references to the SocketChannel named socket with the more generic ByteChannel named channel, which is a superclass of both SocketChannel and FileChannel.

I've added the following methods:

getSerialConnection() - A method used to get a serial connection for communication

getSerialPort() - A method that returns the name of the serial port used for serial communication.  This defaults to "/dev/ttyS0".  On Windows, this should be set to the appropriate comm port (e.g. COM1).  For use with  FTDI serial-to-USB adapters on Linux, use "/dev/ttyUSB0".

setSerialPort(String serialPortName) - A method used to set the name of the serial port used for serial communication.

Lastly, I've modified the main() method to add the shutdown hook immediately after the ctdSource object is created.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@571 9e959af2-1677-4b95-8350-622619c7fe3b