This is a substantial set of changes to the CTDSource class.  The changes are largely added to support bi-directional communication with the CTD.  It enables commands to be sent to the instrument to gather specific metadata fields and to set the real time clock.  I've added the following fields:
sentCommand             - A boolean field indicating if a command has been sent to the instrument
commandPrefix           - The command prefix used to send commands to the instrument
commandSuffix           - The command suffix used to send commands to the instrument 
DEFAULT_SAMPLE_INTERVAL - The default sampling interval in samples per hour (i.e every 4 minutes)
sampleInterval          - The sampling interval in samples per hour
sampleIntervalCommand   - The command used to set the sampling interval on the instrument
startSamplingCommand    - The command used to start the autonomous sampling on the instrument
stopSamplingCommand     - The command used to stop the autonomous sampling on the instrument
setDateTimeCommand      - The command used to set the observation datetime on the instrument
getStatusCommand        - The command get status information from the instrument
getConfigurationCommand - The command get configuration information from the instrument
getCalibrationCommand   - The command get calibration information from the instrument 
getEventsCommand        - The command get event information from the instrument
getHardwareCommand      - The command get hardware information from the instrument
hasMetadata             - A boolean field that indicates if all CTDParser metadata fields are set
samplingIsStopped       - A boolean field that indicates if sampling has been stopped on the instrument
command                 - The command used to query or command the instrument
clockIsSynced           - A boolean stating if the instrument clock is synced to NTP time
clockSyncDate           - The datetime when the instrument clock was last synced to NTP time
DATE_FORMAT             - The date format for the timestamp applied to instrument clock
TZ                      - The timezone used for the sample date
ctdParser               - The instance of the CTD Parser class used to parse CTD output

In support of sending commands to the instrument, I've added the following method:

queryInstrument(String command) - A method that queries the instrument to obtain its ID, returns  a boolean result, true if the command succeeds

I've extensively modified the execute() method.  Instead of only reading data samples from the instrument, it now also queries the instrument for particular metadata fields on instantiation, sets the instrument clock on instantiation, and sets the instrument clock on a daily basis, once per day.

To do so, I've modified the state machine and have added multiple states.  The general sequence is:

state 0: check for existing metadata.  wake the instrument if their are none, otherwise collect data in state 10
state 1: send the stop sampling command
state 2: if not set in the CTDParser, get the instrument status metadata.
state 3: handle the status command response. if not set in the CTDParser, get the instrument configuration metadata.
state 4: handle the configuration command response. if not set in the CTDParser, get the instrument calibration metadata.
state 5: handle the calibration command response. if not set in the CTDParser, get the instrument event metadata.
state 6: handle the events command response. if not set in the CTDParser, get the instrument hardware metadata.
state 7: handle the hardware command response.
state 8: if not synced, sync the instrument clock to NTP time.
state 9: restart the instrument sampling
state 10: listen for incoming samples
state 11: flush new samples to the DT, sync the clock on a daily schedule

Lastly, I've added a finally clause to the try/catch statement in execute() to definitively close the byte channel in the event that it is not closed during normal operations.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@576 9e959af2-1677-4b95-8350-622619c7fe3b