In DataProcessor.m, I've added support for temperature-salinity plots.  I've changed the process() method to now consult the 'currentFigureType' variable in the Configure class.  If it is 'timeSeries', createTimeSeriesFigure() is called.  If it is 'temperatureSalinity', createTSFigure() is called.
I've also changed the signature on createTSFigure(), and it now takes more arguments.  Both figure creation methods now take the 'figureStartDate' argument, and the process() method has been changed to pull this information from the Configure instance (after being set by the calling script).

The createTSFigure() method is now no longer a stub, but will create a TS figure based on the duration of data queried from the DataTurbine.  For instance, data up to one year old are colored gray, the latest weekly data are colored blue, the latest daily data are colored green, and the latest observation is marked with a red circle.

TODO: some of the figure properties are hard-coded (marker sizes, colors, etc.) and need to take advantage of the incoming argument properties.

I've changed the orientation of the figure rectangles to be more widescreen in pixel size.  This may affect export to jpeg, and the jpeg size may need adjustment.

I've added support for 'absolute' time references when querying the DataTurbine.  The code can now take a start date and duration rather than just the most recent data for the given duration.

I've changed the plot to include a solid 0.25 line.

I've changed the title to include the earliest and latest observation times, and have rounded the x axis tick marks to fall on even hours.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@579 9e959af2-1677-4b95-8350-622619c7fe3b