I'm adding in a SerialWriter class that creates a new write thread for writing to the serial port.  It implements the Runnable interface so it can be started in a separate thread than the reading thread. It provides the following fields:
out - The output stream underlying serial communication writes
serialWriter - The buffered writer underlying serial communication writes
writeBuffer - The write byte buffer underlying serial communication writes 
bytesWritten - The number of bytes written to the serial port as an int
isComplete - A boolean field indicating if a write operation is complete
DEFAULT_LOG_CONFIGURATION_FILE - The default log configuration file location
logConfigurationFile - The log configuration file location
logger - The Logger instance used to log system messages

And the following methods:

SerialWriter() - The constructor - creates a new SerialWriter object given an OutputStream to write to
run() -  A method that implements the Runnable interface.  While running, looks for bytes to be written in the writeBuffer ByteBuffer, and writes them to the OutputStream.
write() - A method used to write bytes to the serial port.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@602 9e959af2-1677-4b95-8350-622619c7fe3b