In CTDSource, I've added a boolean field that can be set to allow/disallow sending commands to the CTD instrument:
enableSendCommands - A boolean field that indicates if sending commands should be enabled

I've subsequently changedthe execute() method to use this boolean fied to determine which state to enter.  If it is true, state will be set to 0 and a series of queries will be sent to the CTD to gather metadata.  Otherwise, it will skip to state = 11, currently the state that just starts listening for automated data streaming from the instrument.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@609 9e959af2-1677-4b95-8350-622619c7fe3b