Two methods that deal with command line options:
setArgs() - A method that sets the command line arguments for this class.  This method calls the RBNBSource.setBaseArgs() method.
setOptions() - A method that sets the command line options for this class.  This method calls the RBNBSource.setBaseOptions() method in order to set properties such as the sourceHostName, sourceHostPort, serverName, and serverPort.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@647 9e959af2-1677-4b95-8350-622619c7fe3b