A class used to harvest a decimal ASCII data from a Seacat 16plus CTD) from an emails created by a Satlantic STOR-X logger. The data are converted into RBNB frames and pushed into the RBNB DataTurbine real time server. This class extends org.nees.rbnb.RBNBSource, which in turn extends org.nees.rbnb.RBNBBase, and therefore follows the API conventions found in the org.nees.rbnb code.
The parsing of the data stream relies on the premise that each email contains the text/plain body and an application/octet-stream attachement with the STOR-X binary data format with the embedded ASCII data strings. The strings are parsed out of the binary file using 'SAT' as line beginnings and '\r\n' as endings. Since we don't have a binary format specification for the attachment, other structures in the file are ignored for the time being.

The StorXDispatcher has the following fields:

DEFAULT_ARCHIVE_MODE - A default archive mode for the given source connection to the RBNB server. Valid modes include 'append', 'create', 'load' and 'none'.
archiveMode - The mode in which the source interacts with the RBNB archive. Valid modes include 'append', 'create', 'load' and 'none', however, Kilo Nalu instruments should append to an archive, which will create one if none exist.
DEFAULT_BUFFER_SIZE - The default size of the ByteBuffer for data from the instrument
bufferSize - The size of the ByteBuffer for data from the instrument
DEFAULT_RBNB_CHANNEL - A default RBNB channel name for the given source instrument  
rbnbChannelName - The name of the RBNB channel for this data stream
DEFAULT_SOURCE_HOST_NAME - A default source address for the given source email server
sourceHostName - A source address for the given source email server
DEFAULT_SERVER_NAME - The default IP address or DNS name of the RBNB server
DEFAULT_SERVER_PORT - The default TCP port of the RBNB server
serverName - The IP address or DNS name of the RBNB server
serverPort - The default TCP port of the RBNB server
server - The address and port string for the RBNB server
sampleByteCount - The number of bytes in the ensemble as each byte is read from the stream
DEFAULT_LOG_CONFIGURATION_FILE - The default log configuration file location
logConfigurationFile - The log configuration file location
logger - The Logger instance used to log system messages
xmlConfigurationFile - The XML configuration file location for the list of sensor properties
xmlConfiguration - The XML configuration object with the list of sensor properties
state - The state used to track the data processing
sourceMap - A hash map that contains sensor serial number to RBNB Source mappings
storXParser - The instance of the StorX Parser class used to parse StorX output
startTime - the start time for data
endTime - the end time for data export
executeInterval - The execute interval used to periodically fetch data (in seconds)
connected - a flag to indicate if we are connected to the RBNB server or not
RETRY_INTERVAL - An internal Thread setting used to specify how long, in milliseconds, the execution of the data streaming Thread should wait before re-executing

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@650 9e959af2-1677-4b95-8350-622619c7fe3b