I've added the following method to the StorXDispatcher:
execute() - A method that executes the reading of data from the email account to the RBNB server after all configuration of settings, connections to hosts, and thread initiatizing occurs. This method contains the detailed code for reading the data and interpreting the data files.

This method is called from a Timer object in the main method on a schedule.

git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@651 9e959af2-1677-4b95-8350-622619c7fe3b