I've added the following getter methods:
getBufferSize() - A method that sets the size, in bytes, of the ByteBuffer used in streaming data from a source instrument via a TCP connection
getHostName() - A method that returns the domain name or IP address of the email source (i.e. the IMAP mail server)
getRBNBChannelName() - A method that returns the name of the RBNB channel that contains the streaming data from this instrument
getCVSVersionString() - A method that returns the versioning info for this file.  In this case, it returns a String that includes the Subversion LastChangedDate, LastChangedBy, LastChangedRevision, and HeadURL fields


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@653 9e959af2-1677-4b95-8350-622619c7fe3b