A class that represents a single (full) binary frame from a Satlantic ISUS V3 nitrate concentration instrument. The class represents both light and dark frames, and provides access to the individual fields of the binary data sample as described in the ISUS V3 Operation Manual.
The class includes the following fields:

DEFAULT_LOG_CONFIGURATION_FILE - The default log configuration file location
logConfigurationFile - The log configuration file location
logger - The Logger instance used to log system messages
ISUS_FRAME_SIZE - An ISUS frame size in bytes as an integer
FRAME_DATE_FORMAT - The date format for the timestamp applied to a ISUS frame (Julian day)
TZ - The timezone used for the sample date
isusFrame - An ISUS frame as a ByteBuffer
header - The frame header or synchronization string starts with 'SAT' for a Satlantic instrument, followed by three characters identifying the frame type. The last four characters are the instrument serial number.
sampleDate - The date field denotes the date at the time of the sample, using the year and Julian day. The format is YYYYDDD.
sampleTime - The time field gives the GMT/UTC time of the sample in decimal hours of the day.
nitrogenConcentration - The Nitrate concentration as calculated by the ISUS in μMol/L
auxConcentration1 - The auxiliary 1 fitting result of the ISUS is reported.
auxConcentration2 - The auxiliary 2 fitting result of the ISUS is reported.
auxConcentration3 - The auxiliary 3 fitting result of the ISUS is reported.
rmsError - The Root Mean Square Error of the ISUS’ concentration calculation is given, in ASCII frames to 6 decimal places. 
insideTemperature - The temperature inside the ISUS housing is given in degrees Celsius.
spectrometerTemperature - The temperature of the spectrometer is given in degrees Celsius.
lampTemperature - The temperature of the lamp is given in degrees Celsius.
lampTime - The lamp on-time of the current data acquisition in seconds.
humidity - The humidity inside the instrument, given in percent. Increasing values of humidity indicate a slow leak.
lampVoltage12 - The voltage of the lamp power supply.
internalPowerVoltage5 - The voltage of the internal analog power supply.
mainPowerVoltage - The voltage of the main internal supply.
referenceAverage - The average Reference Channel measurement during the sample time.
referenceVariance - The variance of the Reference Channel measurements.
seaWaterDarkCounts - The Sea-Water Dark calculation in spectrometer counts.
averageWavelength - The average value of all spectrometer channels.
channelWavelengths - The spectrometer counts of the channel wavelengths (256 @ 2 bytes total).
checksum - Binary frames only: A check sum validates binary frames.
timestamp - An ISUS frame timestamp.


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@758 9e959af2-1677-4b95-8350-622619c7fe3b