FRom the previous commit, I've added the following methods:
getHeader() - Return the frame header or synchronization string starts with SAT for a Satlantic instrument, followed by threecharacters identifying the frame type. The last four characters are the instrument serial number.
getSerialNumber() - Get the frame serial number as a String
getSampleDate() - Return the date field denotes the date at the time of the sample, using the year and Julian day. The format is YYYYDDD.
getSampleTime() - Return the time field gives the GMT/UTC time of the sample in decimal hours of the day.
getNitrogenConcentration()  - Return the nitrogen concentration as a float
getAuxConcentration1 - Return the first auxiliary concentration as a float.
getAuxConcentration2() - Return the second auxiliary concentration as a float.
getAuxConcentration3() - Return the third auxiliary concentration as a float.
getRmsError() - Return the Root Mean Square Error of the ISUS’ concentration calculation as a float.
public float {
  return this.rmsError.getFloat();
}

getInsideTemperature() - Return the temperature inside the housing in degrees Celcius as a float.

getSpectrometerTemperature()  - Return the temperature of the spectrometer in degrees Celcius as a float.

getLampTemperature()  - Return the temperature of the lamp in degrees Celcius as a float.

getLampTime()  - Return the lamp on-time of the current data acquisition in seconds as a float.

getHumidity() - Return the humidity inside the instrument, given in percent as a float. Increasing values of humidity indicate a slow leak.

getLampVoltage12() - Return the voltage of the lamp power supply as a float.

getInternalPowerVoltage5() - Return the voltage of the internal analog power supply as a float.

getMainPowerVoltage() - Return the voltage of the main internal supply as a float.

getReferenceAverage() - Return the average Reference Channel measurement during the sample time as a float.

getReferenceVariance()  - Return the variance of the Reference Channel measurements as a float.

getSeaWaterDarkCounts() - Return the Sea-Water Dark calculation in spectrometer counts as a float.

getAverageWavelength() - Return the average value of all spectrometer channels as a float.

getChannelWavelengths()  - Return the counts of a given channel wavelength of the spectrometer as a float.

getChecksum() - Return a check sum to validate binary frames. Satlantic’s software rejects invalid frames.

getTimestamp() - Get the frame timestamp field as a byte array. The timestamp format is YYYYDDD from the first 3 bytes, and HHMMSS.SSS from the last four:
Example:
1E AC CC = 2010316 (year 2010, julian day 316)
09 9D 3E 20 = 16:13:00.000 (4:13 pm)


git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@761 9e959af2-1677-4b95-8350-622619c7fe3b